requirements.txt
boto==2.42.0
dj-database-url==0.4.0
Django==1.9.2
django-allauth==0.28.0
django-bootstrap3==7.1.0
django-contrib-comments==1.7.3
django-gravatar2==1.4.0
django-storages==1.5.1
django-tagging==0.4.5
django-updown==1.0.1
gunicorn==19.4.5
oauthlib==2.0.0
Pillow==3.4.1
psycopg2==2.6.1
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.7.0
whitenoise==2.0.6
